Move experimental sbatch argument adjustment to slurm_utils and make it optional via the adjust_to_partition cookiecutter option.
