+++ {{cookiecutter.profile_name}}/slurm_utils.py
@@ -8,9 +8,9 @@ import subprocess
 from snakemake import io
 
 
-SBATCH_DEFAULTS = "qos=priority" #"""{{cookiecutter.default_arguments}}"""
+SBATCH_DEFAULTS = """{{cookiecutter.default_arguments}}"""
 DEFAULT_CLUSTER_CONFIG = os.path.expandvars("{{cookiecutter.default_cluster_config}}")
-DEFAULT_CLUSTER_CONFIG = os.path.expandvars("$PROJECT_ROOT/slurm_config.yaml")
+ADJUST_TO_PARTIION = bool("{{cookiecutter.adjust_to_partition}}")
 
 RESOURCE_MAPPING = {
         "time": ("time", "runtime", "walltime"),
@@ -78,3 +78,115 @@ def submit_job(jobscript, **sbatch_options):
     except Exception as e:
         raise e
     return jobid
+
+
+def adjust_to_partition(**arg_dict):
+    """Experimental adjustment of sbatch arguments to the given or default partition.
+    """
+    adjusted_args = {}
+    if not ADJUST_TO_PARTIION:
+        return adjusted_args
+
+    partition = arg_dict.get('partition', None) or _get_default_partition()
+    constraint = arg_dict.get("constraint", None)
+    ntasks = int(arg_dict.get("ntasks", 1))
+    nodes = int(arg_dict.get("nodes", 1))
+    mem = arg_dict.get("mem", None)
+    # Determine partition with features. If no constraints have been set,
+    # select the partition with lowest memory
+    try:
+        config = _get_cluster_configuration(partition)
+        mem_feat = _get_features_and_memory(partition)
+        MEMORY_PER_PARTITION = _get_available_memory(mem_feat, constraint)
+        MEMORY_PER_CPU = MEMORY_PER_PARTITION / int(config["cpus"])
+    except Exception as e:
+        print(e)
+        raise e
+    # Adjust memory in the single-node case only; getting the
+    # functionality right for multi-node multi-cpu jobs requires more
+    # development
+    if "nodes" not in arg_dict or nodes == 1:
+        if mem:
+            adjusted_args["mem"] = min(int(mem), MEMORY_PER_PARTITION)
+            AVAILABLE_MEM = ntasks * MEMORY_PER_CPU
+            if adjusted_args["mem"] > AVAILABLE_MEM:
+                adjusted_args["ntasks"] = int(math.ceil(int(mem) / MEMORY_PER_CPU))
+        adjusted_args["ntasks"] = min(int(config["cpus"]), ntasks)
+    else:
+        if nodes == 1:
+            # Allocate at least as many tasks as requested nodes
+            adjusted_args["ntasks"] = nodes
+    # Update time. If requested time is larger than maximum allowed time, reset
+    try:
+        if "time" in arg_dict:
+            adjusted_args["time"] = min(int(config["time"]), int(arg_dict["time"]))
+    except Exception as e:
+        print(e)
+        raise e
+    return adjusted_args
+
+
+def _get_default_partition():
+    """Retrieve default partition for cluster"""
+    res = subprocess.check_output(["sinfo", "-O", "partition"])
+    m = re.search("(?P<partition>\S+)\*", res.decode(), re.M)
+    partition = m.group("partition")
+    return partition
+
+
+def _get_cluster_configuration(partition):
+    """Retrieve cluster configuration for a partition."""
+    # Retrieve partition info; we tacitly assume we only get one response
+    cmd = " ".join(
+        ["sinfo -e -O \"partition,cpus,memory,time,size,maxcpuspernode\"",
+         "-h -p {}".format(partition)])
+    res = subprocess.run(cmd, check=True, shell=True, stdout=subprocess.PIPE)
+    m = re.search("(?P<partition>\S+)\s+(?P<cpus>\d+)\s+(?P<memory>\S+)\s+((?P<days>\d+)-)?(?P<hours>\d+):(?P<minutes>\d+):(?P<seconds>\d+)\s+(?P<size>\S+)\s+(?P<maxcpus>\S+)",
+                  res.stdout.decode())
+    d = m.groupdict()
+    if not 'days' in d or not d['days']:
+        d['days'] = 0
+    d["time"] = int(d['days']) * 24 * 60 + \
+        int(d['hours']) * 60 + int(d['minutes']) + \
+        math.ceil(int(d['seconds']) / 60)
+    return d
+
+
+def _get_features_and_memory(partition):
+    """Retrieve features and memory for a partition in the cluster
+    configuration. """
+    cmd = " ".join(
+        ["sinfo -e -O \"memory,features_act\"",
+         "-h -p {}".format(partition)])
+    res = subprocess.run(cmd, check=True, shell=True, stdout=subprocess.PIPE)
+    mem_feat = []
+    for x in res.stdout.decode().split("\n"):
+        if not re.search("^\d+", x):
+            continue
+        m = re.search("^(?P<mem>\d+)\s+(?P<feat>\S+)", x)
+        mem_feat.append({'mem': m.groupdict()["mem"],
+                         'features': m.groupdict()["feat"].split(",")})
+    return mem_feat
+
+
+def _get_available_memory(mem_feat, constraints=None):
+    """Get available memory
+
+    If constraints are given, parse constraint string into array of
+    constraints and compare them to active features. Currently only
+    handles comma-separated strings and not the more advanced
+    constructs described in the slurm manual.
+
+    Else, the minimum memory for a given partition is returned.
+
+    """
+    if constraints is None:
+        return min([int(x['mem']) for x in mem_feat])
+    try:
+        constraint_set = set(constraints.split(","))
+        for x in mem_feat:
+            if constraint_set.intersection(x["features"]) == constraint_set:
+                return int(x["mem"])
+    except Exception as e:
+        print(e)
+        raise
